# api/openapi/user.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Graph RAG Service - User
  description: >
    Authentication and user accounts. Access tokens are JWTs that carry the user id,
    role, and permissions so other services can authorize requests.
servers:
  - url: https://{hostname}/api
    variables:
      hostname:
        default: localhost
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      operationId: register
      summary: Register a new user account
      description: Command. Creates an account with role user. Returns no body.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '204':
          description: User created
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      operationId: login
      summary: Authenticate and receive an access token
      description: Validates credentials, opens a session, and returns a JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      operationId: logout
      summary: Log out the current session
      description: Command. Revokes the session embedded in the access token.
      responses:
        '204':
          description: Session revoked
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteAccount
      summary: Delete the authenticated user account
      description: Command. Deletes the account and its sessions.
      responses:
        '204':
          description: Account deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    AuthResult:
      type: object
      required: [accessToken, expiresIn, tokenType, user]
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
        tokenType:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      required: [id, email, username, role, permissions, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        permissions:
          type: array
          description: Derived from role. Present in the JWT for other services.
          items:
            type: string
            enum: [knowledge:write, conversation:ask]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      required:
        - slug
        - message
      properties:
        slug:
          type: string
        message:
          type: string