# internal/user/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package ports

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for UserPermissions.
const (
	ConversationAsk UserPermissions = "conversation:ask"
	KnowledgeWrite  UserPermissions = "knowledge:write"
)

// Valid indicates whether the value is a known member of the UserPermissions enum.
func (e UserPermissions) Valid() bool {
	switch e {
	case ConversationAsk:
		return true
	case KnowledgeWrite:
		return true
	default:
		return false
	}
}

// Defines values for UserRole.
const (
	UserRoleAdmin UserRole = "admin"
	UserRoleUser  UserRole = "user"
)

// Valid indicates whether the value is a known member of the UserRole enum.
func (e UserRole) Valid() bool {
	switch e {
	case UserRoleAdmin:
		return true
	case UserRoleUser:
		return true
	default:
		return false
	}
}

// AuthResult defines model for AuthResult.
type AuthResult struct {
	AccessToken string `json:"accessToken"`

	// ExpiresIn Access token lifetime in seconds
	ExpiresIn int `json:"expiresIn"`

	// TokenType Example: Bearer
	TokenType string `json:"tokenType"`
	User      User   `json:"user"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
	Username string              `json:"username"`
}

// User defines model for User.
type User struct {
	CreatedAt time.Time           `json:"createdAt"`
	Email     openapi_types.Email `json:"email"`
	Id        openapi_types.UUID  `json:"id"`

	// Permissions Derived from role. Present in the JWT for other services.
	Permissions []UserPermissions `json:"permissions"`
	Role        UserRole          `json:"role"`
	UpdatedAt   time.Time         `json:"updatedAt"`
	Username    string            `json:"username"`
}

// UserPermissions defines model for User.Permissions.
type UserPermissions string

// UserRole defines model for User.Role.
type UserRole string

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody = RegisterRequest